refactor: 페이지 컴포넌트에서 컴포넌트 분리
